fix(GA): added chromosome type
